+++ server/src/main.rs
@@ -68,7 +68,7 @@ fn main() -> smol::io::Result<()> {
         let mut timecur: f64 = current_time();
         let mut timeprv: f64 = timecur;
         let mut timeacc: f64 = 0.0;
-        let mut tick: u64 = 0;
+        let mut tick: usize = 0;
 
         let mut world = World::new();
