+++ shared/src/systems/mod.rs
@@ -6,6 +6,6 @@ pub use movement::*;
 #[derive(Debug)]
 pub struct Time {
     pub time: f64,
-    pub tick: u64,
+    pub tick: usize,
     pub frame_percent: f64,
 }
